+++ src/main/java/com/demoboletto/dto/request/CreateSpeechDto.java
@@ -22,6 +22,6 @@ public record CreateSpeechDto(
         int rotation,
         @NotNull(message = "scale can not be null")
         @JsonProperty("scale") @Schema(description = "speech scale", example = "0")
-        int scale
+        float scale
 ) {
 }
